feat: Add PokemonApi project
The commit adds the PokemonApi project to the repository. This project includes updates to the npm dependencies and the addition of Angular Material. It also adds environment files and updates styles. Additionally, the commit introduces the PokeService for fetching Pokémon data and adds the CardComponent and GridComponent for displaying Pokémon cards. These components are part of the initial commit of the repository.